[Bronze II] Title: Hashing, Time: 64 ms, Memory: 11516 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15829. Hashing/Hashing.java
@@ -1,23 +1,23 @@
 import java.io.*;
 public class Main {
-    static final int M = 1234567891;
-    static final int R = 31;
+    static final long  M = 1234567891;
+    static final long  R = 31;
     public static void main(String[] args) throws IOException{
         BufferedReader br= new BufferedReader(new InputStreamReader(System.in));
         int N= Integer.parseInt(br.readLine());
         String s = br.readLine();
 
-        int result = 0;
+        long result = 0;
         for(int i = 0; i < N; i++){
             int val = s.charAt(i) - 'a' + 1;         
             result = (result + (val * getPow(i)) % M) % M;  
         }
         System.out.println(result);
     }    
-    public static int getPow(int exp){
+    public static long getPow(int exp){
         if(exp == 1) return R;
         else if(exp == 0) return 1;
-        int div = getPow(exp / 2) % M;
+        long div = getPow(exp / 2) % M;
         if(exp % 2 == 0){
             return (div * div) % M;
         }else{
